[Update] Combind formaturl.js, replace-storage-url-prefix.js, url-transformer.js into url-processor.js
