+++ src/utils/index.js
@@ -1,10 +1,8 @@
-export * from './formaturl'
 export * from './denormalize-articles'
 export * from './detect-device'
 export * from './get-embedded-query-param'
 export * from './get-list-id'
 export * from './get-screen-type'
 export * from './image-processor'
-export * from './replace-storage-url-prefix'
-export * from './url-transformer'
 export * from './check-page-type'
+export * from './url-processor'
