--- src/utils/url-transformer.js
@@ -1,5 +1,0 @@
-import { basePath } from '../constants/index'
-
-export const getAbsPath = (pathname='', search='') => {
-  return basePath + pathname + search
-}
